test: split performance optimizer coverage specs
