+++ tests/unit/performance/PerformanceOptimizer.comprehensive.test.js
@@ -1,0 +1,668 @@
+import fs from 'node:fs';
+import path from 'node:path';
+
+jest.mock('../../../scripts/utils/Logger.js', () => ({
+  __esModule: true,
+  default: {
+    success: jest.fn(),
+    start: jest.fn(),
+    ready: jest.fn(),
+    info: jest.fn(),
+    debug: jest.fn(),
+    warn: jest.fn(),
+    error: jest.fn(),
+    log: jest.fn(),
+  },
+}));
+
+const PERFORMANCE_HTML_FIXTURE = fs.readFileSync(
+  path.resolve(__dirname, '../../mocks/performance/performance-html-fixture.html'),
+  'utf8'
+);
+
+// =========================================
+// 合併自 PerformanceOptimizer.comprehensive.test.js
+// =========================================
+/**
+ * PerformanceOptimizer 全面測試套件
+ * 目標：提升覆蓋率至 80%+
+ */
+
+describe('PerformanceOptimizer - 全面測試', () => {
+  let PerformanceOptimizer = null;
+  let optimizer = null;
+  let mockDocument = null;
+  let originalDocument = null;
+  let originalWindow = null;
+  let originalImage = null;
+  let originalPerformance = null;
+  let originalRequestIdleCallback = null;
+  let originalRequestAnimationFrame = null;
+  let originalLogger = null;
+  let originalChrome = null;
+
+  beforeEach(() => {
+    jest.useFakeTimers();
+    originalDocument = globalThis.document;
+    originalWindow = globalThis.window;
+    originalImage = globalThis.Image;
+    originalPerformance = globalThis.performance;
+    originalRequestIdleCallback = globalThis.requestIdleCallback;
+    originalRequestAnimationFrame = globalThis.requestAnimationFrame;
+    originalLogger = globalThis.Logger;
+    originalChrome = globalThis.chrome;
+    // 創建新的 DOM 環境
+    document.body.innerHTML = PERFORMANCE_HTML_FIXTURE;
+    mockDocument = document;
+
+    // 使用 Jest 原生的 globalThis
+    globalThis.performance = {
+      now: jest.fn(() => Date.now()),
+      memory: {
+        usedJSHeapSize: 10_000_000,
+        totalJSHeapSize: 20_000_000,
+        jsHeapSizeLimit: 100_000_000,
+      },
+    };
+    globalThis.requestIdleCallback = jest.fn(callback => setTimeout(callback, 0));
+    globalThis.requestAnimationFrame = jest.fn(callback => setTimeout(callback, 16));
+
+    // 模擬 Logger
+    globalThis.Logger = {
+      info: jest.fn(),
+      warn: jest.fn(),
+      error: jest.fn(),
+      debug: jest.fn(),
+    };
+
+    // 重新加載模塊
+    jest.resetModules();
+    const module = require('../../../scripts/performance/PerformanceOptimizer');
+    PerformanceOptimizer = module.PerformanceOptimizer;
+
+    optimizer = new PerformanceOptimizer({
+      enableCache: true,
+      enableBatching: true,
+      cacheMaxSize: 10,
+      cacheTTL: 300_000,
+      batchDelay: 16,
+    });
+  });
+
+  afterEach(() => {
+    globalThis.document = originalDocument;
+    globalThis.window = originalWindow;
+    globalThis.Image = originalImage;
+    if (originalPerformance === undefined) {
+      delete globalThis.performance;
+    } else {
+      globalThis.performance = originalPerformance;
+    }
+    if (originalRequestIdleCallback === undefined) {
+      delete globalThis.requestIdleCallback;
+    } else {
+      globalThis.requestIdleCallback = originalRequestIdleCallback;
+    }
+    if (originalRequestAnimationFrame === undefined) {
+      delete globalThis.requestAnimationFrame;
+    } else {
+      globalThis.requestAnimationFrame = originalRequestAnimationFrame;
+    }
+    if (originalLogger === undefined) {
+      delete globalThis.Logger;
+    } else {
+      globalThis.Logger = originalLogger;
+    }
+    if (originalChrome === undefined) {
+      delete globalThis.chrome;
+    } else {
+      globalThis.chrome = originalChrome;
+    }
+    jest.useRealTimers();
+    jest.clearAllMocks();
+  });
+
+  describe('構造函數和初始化', () => {
+    test('應該使用默認選項創建實例', () => {
+      const defaultOptimizer = new PerformanceOptimizer();
+      expect(defaultOptimizer.options.enableCache).toBe(true);
+      expect(defaultOptimizer.options.enableBatching).toBe(true);
+      expect(defaultOptimizer.options.cacheMaxSize).toBe(100);
+    });
+
+    test('應該合併自定義選項', () => {
+      const customOptimizer = new PerformanceOptimizer({
+        cacheMaxSize: 50,
+        batchDelay: 32,
+      });
+      expect(customOptimizer.options.cacheMaxSize).toBe(50);
+      expect(customOptimizer.options.batchDelay).toBe(32);
+      expect(customOptimizer.options.enableCache).toBe(true); // 默認值
+    });
+
+    test('應該初始化緩存統計', () => {
+      expect(optimizer.cacheStats.hits).toBe(0);
+      expect(optimizer.cacheStats.misses).toBe(0);
+      expect(optimizer.cacheStats.evictions).toBe(0);
+      expect(optimizer.cacheStats.prewarms).toBe(0);
+    });
+
+    test('應該初始化批處理統計', () => {
+      expect(optimizer.batchStats.totalBatches).toBe(0);
+      expect(optimizer.batchStats.totalItems).toBe(0);
+      expect(optimizer.batchStats.averageBatchSize).toBe(0);
+    });
+  });
+
+  describe('cachedQuery - DOM 查詢緩存', () => {
+    test('應該執行並緩存查詢結果', () => {
+      const result1 = optimizer.cachedQuery('img', mockDocument);
+      expect(result1).toBeDefined();
+      expect(result1.length).toBeGreaterThan(0);
+      expect(optimizer.cacheStats.misses).toBe(1);
+      expect(optimizer.queryCache.size).toBe(1);
+    });
+
+    test('應該從緩存返回結果', () => {
+      // 創建全新的 optimizer 實例以避免測試間的干擾
+      const freshOptimizer = new PerformanceOptimizer({
+        enableCache: true,
+      });
+
+      const result1 = freshOptimizer.cachedQuery('img', mockDocument);
+      const result2 = freshOptimizer.cachedQuery('img', mockDocument);
+
+      // 驗證緩存命中統計 - 第二次查詢應該命中緩存
+      expect(freshOptimizer.cacheStats.hits).toBeGreaterThan(0);
+      // 兩次查詢應該返回相同數量的元素
+      expect(result2).toHaveLength(result1.length);
+    });
+
+    test('應該支持 single 選項', () => {
+      const result = optimizer.cachedQuery('img', mockDocument, { single: true });
+      expect(result).toBeDefined();
+      expect(result.nodeType).toBe(1); // 單個元素
+    });
+
+    test('應該支持 all 選項', () => {
+      const result = optimizer.cachedQuery('img', mockDocument, { all: true });
+      expect(result).toBeDefined();
+      expect(result.length).toBeGreaterThan(0);
+    });
+
+    test('應該處理無效選擇器', () => {
+      const beforeSize = optimizer.queryCache.size;
+      const beforeKeys = Array.from(optimizer.queryCache.keys());
+      const result = optimizer.cachedQuery(':::invalid:::', mockDocument);
+      expect(result).toEqual([]);
+      expect(optimizer.queryCache.size).toBe(beforeSize);
+      expect(Array.from(optimizer.queryCache.keys())).toEqual(beforeKeys);
+    });
+
+    test('應該在禁用緩存時直接查詢', () => {
+      optimizer.options.enableCache = false;
+      const result = optimizer.cachedQuery('img', mockDocument);
+      expect(result).toBeDefined();
+      expect(optimizer.queryCache.size).toBe(0);
+    });
+
+    test('應該驗證緩存的元素是否仍在 DOM 中', () => {
+      // 簡化測試，只驗證緩存邏輯，不涉及實際 DOM 操作
+      optimizer.cachedQuery('p', mockDocument);
+      expect(optimizer.queryCache.size).toBe(1);
+
+      // 第二次查詢應該命中緩存
+      optimizer.cachedQuery('p', mockDocument);
+      expect(optimizer.cacheStats.hits).toBeGreaterThan(0);
+    });
+
+    test('應該在相同 selector 且不同 context 時生成不同的緩存鍵', () => {
+      const contextA = mockDocument.createElement('div');
+      contextA.id = 'ctx-a';
+      contextA.innerHTML = '<article><p>A</p></article>';
+
+      const contextB = mockDocument.createElement('section');
+      contextB.id = 'ctx-b';
+      contextB.innerHTML = '<article><p>B</p></article>';
+
+      mockDocument.body.append(contextA, contextB);
+
+      optimizer.cachedQuery('article', contextA);
+      optimizer.cachedQuery('article', contextB);
+
+      expect(optimizer.queryCache.size).toBe(2);
+      expect(new Set(Array.from(optimizer.queryCache.keys())).size).toBe(2);
+    });
+  });
+
+  describe('clearExpiredCache - 過期緩存清理', () => {
+    test('應該清理過期的緩存', () => {
+      // 添加一些緩存
+      optimizer.cachedQuery('img', mockDocument);
+      optimizer.cachedQuery('p', mockDocument);
+
+      // 修改緩存時間戳
+      const firstKey = optimizer.queryCache.keys().next().value;
+      const firstEntry = optimizer.queryCache.get(firstKey);
+      firstEntry.timestamp = Date.now() - 400_000; // 過期
+
+      const clearedCount = optimizer.clearExpiredCache({ maxAge: 300_000 });
+      expect(clearedCount).toBe(1);
+    });
+
+    test('應該支持強制清理所有緩存', () => {
+      optimizer.cachedQuery('img', mockDocument);
+      optimizer.cachedQuery('p', mockDocument);
+
+      const clearedCount = optimizer.clearExpiredCache({ force: true });
+      expect(clearedCount).toBe(2);
+      expect(optimizer.queryCache.size).toBe(0);
+    });
+
+    test('應該使用默認的 cacheTTL', () => {
+      optimizer.cachedQuery('img', mockDocument);
+      const firstKey = optimizer.queryCache.keys().next().value;
+      const firstEntry = optimizer.queryCache.get(firstKey);
+      firstEntry.timestamp = Date.now() - optimizer.options.cacheTTL - 1000;
+
+      const clearedCount = optimizer.clearExpiredCache();
+      expect(clearedCount).toBe(1);
+    });
+  });
+
+  describe('refreshCache - 刷新緩存', () => {
+    test('應該刷新單個選擇器的緩存', () => {
+      optimizer.cachedQuery('img', mockDocument);
+
+      // 獲取緩存鍵（可能不是簡單的字符串）
+      const cacheKeys = Array.from(optimizer.queryCache.keys());
+      expect(cacheKeys).toHaveLength(1);
+
+      const firstResult = optimizer.queryCache.get(cacheKeys[0]);
+      const firstTimestamp = firstResult.timestamp;
+
+      // 稍微延遲以確保時間戳不同
+      jest.advanceTimersByTime(10);
+
+      // 刷新緩存
+      optimizer.refreshCache('img', mockDocument);
+
+      const refreshedResult = optimizer.queryCache.get(cacheKeys[0]);
+      expect(refreshedResult).toBeDefined();
+      expect(refreshedResult.timestamp).toBeGreaterThanOrEqual(firstTimestamp);
+    });
+
+    test('應該刷新多個選擇器的緩存', () => {
+      optimizer.cachedQuery('img', mockDocument);
+      optimizer.cachedQuery('p', mockDocument);
+
+      optimizer.refreshCache(['img', 'p'], mockDocument);
+
+      expect(optimizer.queryCache.size).toBe(2);
+    });
+
+    test('應該刪除查詢結果為空的緩存', () => {
+      optimizer.cachedQuery('img', mockDocument);
+      const imgCacheKeysBefore = Array.from(optimizer.queryCache.entries())
+        .filter(([_key, value]) => value?.selector === 'img')
+        .map(([key]) => key);
+      expect(imgCacheKeysBefore.length).toBeGreaterThan(0);
+
+      // 移除所有圖片
+      const images = mockDocument.querySelectorAll('img');
+      for (const img of images) {
+        img.remove();
+      }
+
+      optimizer.refreshCache('img', mockDocument);
+
+      // 緩存應該被刪除
+      imgCacheKeysBefore.forEach(key => {
+        expect(optimizer.queryCache.has(key)).toBe(false);
+      });
+      expect(
+        Array.from(optimizer.queryCache.values()).some(cacheEntry => cacheEntry?.selector === 'img')
+      ).toBe(false);
+    });
+  });
+
+  describe('preloadSelectors - 選擇器預熱', () => {
+    test('應該預熱選擇器並緩存結果', async () => {
+      const selectors = ['img', 'p', 'article'];
+      const results = await optimizer.preloadSelectors(selectors, mockDocument);
+
+      expect(results.length).toBeGreaterThan(0);
+      expect(optimizer.cacheStats.prewarms).toBeGreaterThan(0);
+      expect(optimizer.prewarmedSelectors.size).toBeGreaterThan(0);
+    });
+
+    test('應該跳過已預熱的選擇器', async () => {
+      const selectors = ['img', 'p'];
+      await optimizer.preloadSelectors(selectors, mockDocument);
+
+      const prewarmCount1 = optimizer.cacheStats.prewarms;
+
+      // 再次預熱相同的選擇器
+      await optimizer.preloadSelectors(selectors, mockDocument);
+
+      const prewarmCount2 = optimizer.cacheStats.prewarms;
+      expect(prewarmCount2).toBe(prewarmCount1); // 應該跳過
+    });
+
+    test('應該處理預熱失敗的選擇器', async () => {
+      const selectors = [':::invalid:::', 'img'];
+      const results = await optimizer.preloadSelectors(selectors, mockDocument);
+
+      // 至少有一個選擇器成功
+      expect(results.length).toBeGreaterThan(0);
+      // 驗證有成功的緩存
+      const successfulResults = results.filter(result => result.cached);
+      expect(successfulResults.length).toBeGreaterThan(0);
+    });
+
+    test('應該在禁用緩存時返回空數組', async () => {
+      optimizer.options.enableCache = false;
+      const results = await optimizer.preloadSelectors(['img'], mockDocument);
+      expect(results).toEqual([]);
+    });
+
+    test('應該處理無效的選擇器參數', async () => {
+      const results = await optimizer.preloadSelectors(null, mockDocument);
+      expect(results).toEqual([]);
+    });
+  });
+
+  describe('smartPrewarm - 智能預熱', () => {
+    test('應該基於頁面結構預熱選擇器', async () => {
+      const results = await optimizer.smartPrewarm(mockDocument);
+
+      expect(results.length).toBeGreaterThan(0);
+      expect(optimizer.prewarmedSelectors.size).toBeGreaterThan(0);
+    });
+
+    test('應該識別 article 結構', async () => {
+      await optimizer.smartPrewarm(mockDocument);
+
+      // 應該包含 article 相關的選擇器
+      expect(optimizer.queryCache.size).toBeGreaterThan(0);
+    });
+
+    test('應該識別 role="main" 結構', async () => {
+      await optimizer.smartPrewarm(mockDocument);
+
+      // 應該包含 main 相關的選擇器
+      const hasMainRole = Array.from(optimizer.queryCache.keys()).some(key => key.includes('main'));
+      expect(hasMainRole).toBe(true);
+    });
+
+    test('應該識別 CMS 類名模式', async () => {
+      await optimizer.smartPrewarm(mockDocument);
+
+      // 應該包含 .entry-content 相關的選擇器
+      const hasCmsPattern = Array.from(optimizer.queryCache.keys()).some(key =>
+        key.includes('.entry-content')
+      );
+      expect(hasCmsPattern).toBe(true);
+    });
+  });
+
+  describe('batchProcessImages - 批處理圖片', () => {
+    test('應該批處理圖片並返回結果', async () => {
+      const images = [{ src: 'test1.jpg' }, { src: 'test2.jpg' }, { src: 'test3.jpg' }];
+      const processor = img => ({ url: img.src, processed: true });
+
+      const promise = optimizer.batchProcessImages(images, processor);
+
+      // 執行批處理
+      jest.runAllTimers();
+
+      const results = await promise;
+      expect(results).toHaveLength(3);
+      expect(results[0]).toEqual({ url: 'test1.jpg', processed: true });
+    });
+
+    test('應該在禁用批處理時直接處理', async () => {
+      optimizer.options.enableBatching = false;
+      const images = [{ src: 'test.jpg' }];
+      const processor = img => ({ url: img.src });
+
+      const results = await optimizer.batchProcessImages(images, processor);
+      expect(results).toHaveLength(1);
+      expect(results[0]).toEqual({ url: 'test.jpg' });
+    });
+  });
+
+  describe('clearCache - 清理緩存', () => {
+    test('應該強制清理所有緩存', () => {
+      optimizer.cachedQuery('img', mockDocument);
+      optimizer.cachedQuery('p', mockDocument);
+
+      optimizer.clearCache({ force: true });
+      expect(optimizer.queryCache.size).toBe(0);
+    });
+
+    test('應該清理過期的緩存', () => {
+      optimizer.cachedQuery('img', mockDocument);
+
+      // 修改時間戳
+      const key = optimizer.queryCache.keys().next().value;
+      const entry = optimizer.queryCache.get(key);
+      entry.timestamp = Date.now() - 400_000;
+
+      optimizer.clearCache({ maxAge: 300_000 });
+      expect(optimizer.queryCache.size).toBe(0);
+    });
+
+    test('應該使用默認的最大年齡', () => {
+      optimizer.cachedQuery('img', mockDocument);
+
+      const key = optimizer.queryCache.keys().next().value;
+      const entry = optimizer.queryCache.get(key);
+      entry.timestamp = Date.now() - 400_000;
+
+      optimizer.clearCache();
+      expect(optimizer.queryCache.size).toBe(0);
+    });
+  });
+
+  describe('getStats - 獲取統計信息', () => {
+    test('應該返回完整的統計信息', () => {
+      // 直接操作統計，避免 DOM 查詢問題
+      optimizer.cacheStats.hits = 1;
+      optimizer.cacheStats.misses = 1;
+
+      const stats = optimizer.getStats();
+
+      expect(stats.cache).toBeDefined();
+      expect(stats.cache.hits).toBe(1);
+      expect(stats.cache.misses).toBe(1);
+      expect(stats.cache.hitRate).toBeCloseTo(0.5);
+      expect(stats.batch).toBeDefined();
+      expect(stats.metrics).toBeDefined();
+      expect(stats.memory).toBeDefined();
+    });
+
+    test('應該計算緩存命中率', () => {
+      optimizer.cacheStats.hits = 2;
+      optimizer.cacheStats.misses = 1;
+
+      const stats = optimizer.getStats();
+      expect(stats.cache.hitRate).toBeCloseTo(0.67, 1);
+    });
+
+    test('應該返回內存統計或 null', () => {
+      const stats = optimizer.getStats();
+
+      // 內存統計可能為 null（取決於環境）
+      if (stats.memory) {
+        expect(stats.memory.usedJSHeapSize).toBeDefined();
+        expect(stats.memory.totalJSHeapSize).toBeDefined();
+      } else {
+        expect(stats.memory).toBeNull();
+      }
+    });
+  });
+
+  describe('getPerformanceStats - 別名方法', () => {
+    test('應該返回與 getStats 相同的結果', () => {
+      const stats1 = optimizer.getStats();
+      const stats2 = optimizer.getPerformanceStats();
+
+      expect(stats2).toEqual(stats1);
+    });
+  });
+
+  describe('resetStats - 重置統計', () => {
+    test('應該重置所有統計信息', () => {
+      optimizer.cachedQuery('img', mockDocument);
+      optimizer.cachedQuery('p', mockDocument);
+
+      optimizer.resetStats();
+
+      expect(optimizer.cacheStats.hits).toBe(0);
+      expect(optimizer.cacheStats.misses).toBe(0);
+      expect(optimizer.batchStats.totalBatches).toBe(0);
+      expect(optimizer.metrics.domQueries).toBe(0);
+    });
+  });
+
+  describe('_maintainCacheSizeLimit - 維護緩存大小', () => {
+    test('應該在達到限制時移除最舊的項目', () => {
+      // 建立可命中的 selector，確保查詢結果會被快取
+      document.body.innerHTML = Array.from(
+        { length: optimizer.options.cacheMaxSize + 1 },
+        (_unused, i) => `<div id="cache-item-${i}"></div>`
+      ).join('');
+
+      // 填滿緩存
+      for (let i = 0; i < optimizer.options.cacheMaxSize; i++) {
+        optimizer.cachedQuery(`#cache-item-${i}`, mockDocument, { single: true });
+      }
+
+      // 添加新項目
+      optimizer.cachedQuery(`#cache-item-${optimizer.options.cacheMaxSize}`, mockDocument, {
+        single: true,
+      });
+
+      // 大小應該不超過限制
+      expect(optimizer.queryCache.size).toBeLessThanOrEqual(optimizer.options.cacheMaxSize);
+      expect(optimizer.cacheStats.evictions).toBeGreaterThan(0);
+    });
+  });
+
+  describe('便捷函數', () => {
+    test('cachedQuery 函數應該使用默認實例', () => {
+      const { cachedQuery } = require('../../../scripts/performance/PerformanceOptimizer');
+      const result = cachedQuery('img', mockDocument);
+      expect(result).toBeDefined();
+    });
+
+    test('batchProcess 函數應該使用默認實例', async () => {
+      const { batchProcess } = require('../../../scripts/performance/PerformanceOptimizer');
+      const items = [{ id: 1 }];
+      const processor = item => ({ ...item, processed: true });
+
+      const promise = batchProcess(items, processor);
+      jest.runAllTimers();
+
+      const results = await promise;
+      expect(results).toHaveLength(1);
+    });
+  });
+
+  describe('模塊導出', () => {
+    test('應該正確導出到 module.exports', () => {
+      // 驗證 module.exports 導出
+      const exported = require('../../../scripts/performance/PerformanceOptimizer');
+
+      expect(exported.PerformanceOptimizer).toBeDefined();
+      expect(exported.cachedQuery).toBeDefined();
+      expect(exported.batchProcess).toBeDefined();
+    });
+  });
+
+  describe('batchProcessWithRetry helper', () => {
+    test.each([
+      {
+        name: '第一次嘗試就成功並返回結果',
+        items: [1, 2],
+        urlPrefix: 'image',
+        expectedCalls: 1,
+        expectedAttempts: 1,
+        options: {},
+        createCustomBatchFn: () =>
+          jest.fn((items, processor) => Promise.resolve(items.map(item => processor(item)))),
+      },
+      {
+        name: '失敗後重試並最終成功',
+        items: [1, 2, 3],
+        urlPrefix: 'ix',
+        expectedCalls: 2,
+        expectedAttempts: 2,
+        options: {
+          maxAttempts: 3,
+          baseDelay: 0,
+        },
+        createCustomBatchFn: () => {
+          let attempt = 0;
+          return jest.fn((items, processor) => {
+            attempt++;
+            if (attempt === 1) {
+              return Promise.reject(new Error('Batch boom'));
+            }
+            return Promise.resolve(items.map(item => processor(item)));
+          });
+        },
+      },
+    ])('應該在$name', async scenario => {
+      const module = require('../../../scripts/performance/PerformanceOptimizer');
+      const customBatchFn = scenario.createCustomBatchFn();
+      const processor = value => ({ url: `${scenario.urlPrefix}-${value}` });
+
+      const { results, meta } = await module.batchProcessWithRetry(scenario.items, processor, {
+        customBatchFn,
+        captureFailedResults: true,
+        isResultSuccessful: result => Boolean(result?.url),
+        ...scenario.options,
+      });
+
+      expect(customBatchFn).toHaveBeenCalledTimes(scenario.expectedCalls);
+      expect(results).toHaveLength(scenario.items.length);
+      expect(meta.attempts).toBe(scenario.expectedAttempts);
+      expect(meta.failedIndices).toEqual([]);
+    });
+
+    test('當達到最大嘗試次數仍失敗時應返回 null', async () => {
+      const module = require('../../../scripts/performance/PerformanceOptimizer');
+      const customBatchFn = jest.fn(() => Promise.reject(new Error('permanent failure')));
+      const processor = value => ({ url: `img-${value}` });
+
+      const { results, meta } = await module.batchProcessWithRetry([1], processor, {
+        customBatchFn,
+        maxAttempts: 2,
+        baseDelay: 0,
+      });
+
+      expect(customBatchFn).toHaveBeenCalledTimes(2);
+      expect(results).toBeNull();
+      expect(meta.attempts).toBe(2);
+      expect(meta.lastError).toBeInstanceOf(Error);
+    });
+
+    test('應該在 captureFailedResults 時回報失敗索引', async () => {
+      const module = require('../../../scripts/performance/PerformanceOptimizer');
+      const customBatchFn = jest.fn((items, processor) =>
+        Promise.resolve(items.map(item => processor(item)))
+      );
+      const processor = value => (value % 2 === 0 ? null : { url: `valid-${value}` });
+
+      const { meta } = await module.batchProcessWithRetry([1, 2, 3], processor, {
+        customBatchFn,
+        captureFailedResults: true,
+        isResultSuccessful: result => Boolean(result?.url),
+      });
+
+      expect(meta.failedIndices).toEqual([1]);
+    });
+  });
+});
